Fix a few type mismatches with recent gnustep-base.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileFoundation@5940 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Source/ETCollection.m
@@ -203,7 +203,7 @@ NSCountedSet is always mutable and has not immutable equivalent. */
 	NSMutableArray *indexes = [NSMutableArray arrayWithCapacity: [self count]];
 	int nbOfIndexes = [self count];
 	int nbOfCopiedIndexes = -1;
-	unsigned int *copiedIndexes = calloc(sizeof(unsigned int), nbOfIndexes);
+	NSUInteger *copiedIndexes = calloc(sizeof(NSUInteger), nbOfIndexes);
 	
 	nbOfCopiedIndexes = [self getIndexes: copiedIndexes maxCount: nbOfIndexes
 		inIndexRange: NULL];
